+++ Sources/StreamChat/ChatClientFactory.swift
@@ -133,6 +133,7 @@ class ChatClientFactory {
             UserWatchingEventMiddleware(),
             UserUpdateMiddleware(),
             ChannelVisibilityEventMiddleware(),
+            MessageVisibilityEventMiddleware(),
             EventDTOConverterMiddleware()
         ]